% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saudiregions}
\alias{saudiregions}
\title{Saudi Regions, Cities and Districts}
\format{A data frame with 2087 rows and 7 variables: each row is a district.
\describe{
  \item{id}{a unique id}
  \item{region.name}{The name of the region, 13 regions}
  \item{city.name}{City name}
  \item{district.name}{The District Name}
  \item{direction.name}{}
  \item{direction.name}{}
  \item{city.id}{}
  \item{direction.id}{}
}}
\source{
\url{https://github.com/usefksa/Saudi_GIS_Data/}
}
\usage{
saudiregions
}
\description{
A dataset containing all Saudi regions with its cities and districts.
}
\keyword{datasets}
